# Create custom 4x4 and 2x2 linalg operations (matmul, kron, determinant)